+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 -   Update minimum Java to Java 8
 -   Add in Discord rich presence integration
 -   Fix post launch error check dialog not being displayed correctly
+-   Fix fabric not installing correctly